requirements.txt
django
djangorestframework
drf_yasg
